Dev : make token approval utils great again
