+++ packages/arb-token-bridge-ui/src/token-bridge-sdk/BridgeTransferStarter.ts
@@ -1,7 +1,4 @@
 import { Provider } from '@ethersproject/providers'
-import { MaxUint256 } from '@ethersproject/constants'
-import { ERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/ERC20__factory'
-import { Erc20Bridger } from '@arbitrum/sdk'
 import { ERC20BridgeToken } from '../hooks/arbTokenBridge.types'
 import { MergedTransaction } from '../state/app/state'
 import { BigNumber, Signer } from 'ethers'
@@ -42,7 +39,7 @@ export type BridgeTransfer = {
 export type BridgeTransferStarterProps = {
   sourceChainProvider: Provider
   destinationChainProvider: Provider
-  selectedToken: SelectedToken | null
+  selectedToken?: SelectedToken | null
 }
 
 export type RequiresNativeCurrencyApprovalProps = {
@@ -71,11 +68,9 @@ export type ApproveTokenProps = {
   destinationChainProvider: Provider
 }
 
-export type ApproveTokenGasEstimationProps = {
-  erc20L1Address: string
-  address: string
-  l1Provider: Provider
-  l2Provider: Provider
+export type ApproveTokenEstimateGasProps = {
+  amount: BigNumber
+  signer: Signer
 }
 
 export abstract class BridgeTransferStarter {
@@ -88,7 +83,7 @@ export abstract class BridgeTransferStarter {
   constructor(props: BridgeTransferStarterProps) {
     this.sourceChainProvider = props.sourceChainProvider
     this.destinationChainProvider = props.destinationChainProvider
-    this.selectedToken = props.selectedToken
+    this.selectedToken = props.selectedToken ?? null
   }
 
   public abstract requiresNativeCurrencyApproval(
@@ -103,24 +98,9 @@ export abstract class BridgeTransferStarter {
     props: RequiresTokenApprovalProps
   ): Promise<boolean>
 
-  public static async approveTokenGasEstimation({
-    erc20L1Address,
-    address,
-    l1Provider,
-    l2Provider
-  }: ApproveTokenGasEstimationProps) {
-    const erc20Bridger = await Erc20Bridger.fromProvider(l2Provider)
-    const l1GatewayAddress = await erc20Bridger.getL1GatewayAddress(
-      erc20L1Address,
-      l1Provider
-    )
-
-    const contract = ERC20__factory.connect(erc20L1Address, l1Provider)
-
-    return contract.estimateGas.approve(l1GatewayAddress, MaxUint256, {
-      from: address
-    })
-  }
+  public abstract approveTokenEstimateGas(
+    props: ApproveTokenEstimateGasProps
+  ): Promise<BigNumber>
 
   public abstract approveToken(props: ApproveTokenProps): Promise<void>
